[BUGFIX release] Fix toString test to work with Ember 3.2+.
This test was attempting to ensure that the records guid was included in
the `record.toString()` output, but changes in Ember 3.2+ changed the
default output of the class / constructor name portion (therefore
failing the assertion in a useless way).

This updates the test to add a custom `toString` to the model class so
that we can be decoupled from how Ember `.toString()`'s classes when
there is no custom `.toString` setup (because this is actually fairly
unstable version over version).