Replace ember-cli-eslint with direct eslint usage.
`yarn lint:js` is used by CI to enforce linting, removing
ember-cli-eslint effectively means that no linting tests will be done as
part of the build.

This works around a long standing issue in ember-cli-eslint where the
directory provided to eslint when linting is incorrect, which throws an
error when `eslint-config-prettier` attempts to read the `.editorconfig`
file to infer base settings.