Refactor argument parsing for --instrument and --record-data-rfc-build
This simplifies and reorganizes the process.argv processing that needs
to be done for the two flags that we support.

This commit migrates the instrument flag handling to a single file
(`lib/heimdall-utils.js`) and uses that single exported function
throughout the rest of the code. It also removes the reliance on
`process.env` string coercion (which apparently _can_ be broken when
you've reassigned `process.env`).

It also changes around the details of the previous
`--record-data-rfc-build` flag. Instead, we add a
`--disable-record-data-rfc-build` flag (because as of 3.5 beta's the
record data code is enabled by default). During this refactor, the
`config` hook was migrated into the dummy app itself, so that ember-data
no longer adds that flag to all consuming app's runtime config.