Fix tests execution (#1)
* Fix lint error

* Fix `_requireBuildPackages` error

* Reset yarn.lock from master

There seems to be some issues in the yarn.lock

To solve them I ran:
- `git checkout master yarn.lock`
- `yarn`

Throw an error in case of dependency mismatch (#2)

* Throw an error in case of dependency mismatch

An app (or addon) that consumes ember-data should have:
- either jquery
- or ember-fetch version 6.0.0 or above

Throw a build error if it's not the case.

* Remove ember-fetch as a peerDependency

Move ember-fetch to dependencies (#3)

Indeed, two `ember-fetch`'s modules are imported in rest-adapter.