Set the json:api content-type only when needed
When the request is empty, the header wouldn't be needed. Here the
json:api spec regarding the requests headers:
https://github.com/json-api/json-api/blob/ff35bb226ab96260cf0e80aedd87fe84ecb9cab4/_format/1.1/index.md#-universal-responsibilities

This would fix a bug where `contentType` was set on undefined `options`.

I noticed the bug was occuring on `DELETE` requests.

An alternative bug fix would have been:
```javascript
if (options) {
  options.contentType = ...
}
```