Show transform usage after example
Reading the API information about transforms, the order seemed a little off. First, we indicate a basic example for transforms, and the usage that makes reference (`app/transforms/temperature`) is at the end of this section, after an option slightly more advanced.

This change moves the usage snippet right after the example.