Remove Publish job from CI (#5974)
In the past, this job has been used to publish to Bower registry and to
S3. The S3 build has been used to build the doc at some time. Now,
ember-data doc is fetched from npm release to build emberjs guides.

As this job is not publishing anything anymore, it sounds safe to remove
it.