[BUGFIX publish] add script to handle release logic, work around OTP issues"
* Ensures that `yarn build:production` is run for each "addon" package
* Separates out versioning from packaging and publishing
* Updates the RELEASE documentation

We require 2fa to publish ember-data, but 2fa tokens expire after 30s.
`lerna publish` which handles versioning, packaging, and publishing
does not itself accept an `OTP` for 2fa and when passed as an environment
variable it is likely to have long expired prior to reaching the publish
step.

Our script separates out these steps so that we can recover from publishing
errors more gracefully, and surfaces the OTP prompt (and reprompt) at
the correct point in time for the publish.

The opportunity was taken here to add guard rails around many of the finer
points of our release process. Outside of branching that occurs at the
beginning of the release cycle and changelog updates our release process
is not mostly automated.