[BUGFIX release] Remove ember-fetch dependency.
This change removes ember-fetch as a direct dependency of ember-data,
and allows for this dependency to be provided by the host application
or for users to directly use native fetch.

Having `ember-fetch` as a direct dependency of ember-data has proven
troublesome. There are issues using ember-fetch as a nested dependency
(these are documented in the ember-fetch README), but it is also quite
difficult for ember-data to ensure that the versions of `ember-fetch`
provided will actually include the correct version (duplicated addon
merging in the ember-cli addon space is complicated and error prone).

This change moves the `determineBodyPromise` and `serializeQueryParams`
helper functions directly into `@ember-data/adapter`. A future version
of ember-fetch will drop these utilities (as well as the ember-data
adapter mixin). In addition, this also enables usage of the global
`fetch` if a `fetch` module is not provided.