Skip linting for release and lts branches
This updates the Azure pipeline config so that the lint job is skipped
unless the pull request target or the branch name is `master` or `beta`.

This helps to bring the Azure pipeline config closer to feature pairity with
[the config for Travis CI](https://github.com/emberjs/data/blob/2189024320ba10c1086db3aef01c0acdfb1f78f9/.travis.yml#L41).

I took a slightly different approach here by whitelisting the `master`
and `beta` branches rather than trying to blacklist the many other cases
where we want to skip the linting because this is easier to express
using [the syntax for expressions available in Azure](https://docs.microsoft.com/en-us/azure/devops/pipelines/process/expressions?view=azure-devops).

See
https://dev.azure.com/heroiceric/Ember%20Data/_build/results?buildId=4
for an example of a build running for the `master` branch.

See https://dev.azure.com/heroiceric/Ember%20Data/_build/results?buildId=3
for an example of a pull request build running that targets the `master`
branch.

See https://dev.azure.com/ember-data/emberjs/_build/results?buildId=130
for an example of a build that targets the `release` branch where the
`Lint` job is skipped.