[BUGFIX release] Fix fetch detection.
The prior detection of when to use fetch vs when to use jQuery was not
properly acounting for the `fastboot` case when `ember-fetch` is not
present and a `fetch` global is not present. In that scenario, `najax`
would be present (provided by `fastboot` itself).

We already had the correct logic in `_ajax`, however the `useFetch`
computed was returning `true` which prevented the `najax` supported
branch from being used.