Add a CI job for ember-release with jQuery (#5975)
* Add a CI job for ember-release with jQuery

With the recent switch from Query.ajax to fetch in the adapters code
and the jQuery removal in Ember, this job aims at adding a small safety
net.

* Add azure pipelines job
* Ensure @ember/optional-features is present (otherwise setting
  `EMBER_OPTIONAL_FEATURES` has no effect)
* Run tests without jQuery by default
* tests: Merge fetch and ajax adapters tests

From now on, dedicated CI jobs will run with and without jQuery.

This lets us simplify the adapters tests a bit. Indeed, the same tests will run
for fetch (without jQuery) and ajax (with jQuery).

* Refactor rest-adapter integration test...

...to run with both fetch and ajax. Previously, jQuery was always loaded
in the test suite. So this file was not run with fetch. Now, it is run
by default with fetch (not ajax). In CI, both scenarios should be run.