[BUGFIX release beta canary] Fix Model lifecycle event deprecations
Related RFC: https://github.com/emberjs/rfcs/blob/master/text/0329-deprecated-ember-evented-in-ember-data.md
Original PR: https://github.com/emberjs/data/pull/6059

Deprecations were added for the following model lifecycle events
`becameError`, `becameInvalid`, `didCreate`, `didDelete`, `didLoad`,
`didUpdate`, `ready`, `rolledBack` in https://github.com/emberjs/data/pull/6059.

There was a discussion on the PR that suggested adding an additional
check to make sure that that the deprecations were only triggered for
instances of Ember Data's `Model` class https://github.com/emberjs/data/pull/6059#discussion_r284921979

This extra check was added but it accidentally was checking for `this`
to be an instance of the `Model` class, but that can never be true since
`this` will always be an instance of `InternalModel`:

https://github.com/emberjs/data/blob/5e495394cbf67025be2622f604e67004d79bc21c/packages/store/addon/-private/system/model/internal-model.ts#L446-L455

Since we only want to check for these deprecations for instances of
`Model`, I moved the deprecation logic to the `Model` class

https://github.com/emberjs/data/pull/6059 also included a mechanism for
tracking deprecations that were logged per model class so we could
prevent logging multiple deprecations for the same model class and
deprecated lifecycle method pair:

https://github.com/emberjs/data/blob/5e495394cbf67025be2622f604e67004d79bc21c/packages/store/addon/-private/system/model/internal-model.ts#L61-L77

This also fixes a bug with the deprecation tracking logic which was that
we were never actually adding the logged deprecations to the `WeakMap`
meant to keep track of them.