Use GitHub Actions for CI
This supersedes https://github.com/emberjs/data/pull/6098 but with
[GitHub
actions](https://help.github.com/en/articles/about-github-actions)
rather than Azure Pipelines.

I think we should continue running Travis CI alongside this for a little
while to make sure everything is working correctly but we should remove
the Azure Pipelines config.

Why?

- This GitHub version runs much faster because we're given more
  resources to run jobs concurrently.
- The syntax is also nicer, which is my opinion of course but it seems
  to be shared by others.

Stages

The jobs are configured to run in three stages, each only running once
all of the jobs in the previous stage have passed:

1. lint & basic-tests
1. floating-dependencies, lts, releases, additional-scenarios
1. external-partners

This setup is intended to make sure that one build is not using all of
the repo's allowed jobs. It is a trade off between speed and resource
availability that we might want to tweak in the future.

Allowed failures

GitHub actions, like Azure Pipelines, doesn't provide a feature like
[Travis CI's
`allow_failures`](https://docs.travis-ci.com/user/customizing-the-build/#rows-that-are-allowed-to-fail),
which allows a job to fail without failing the entire build.  As a
workaround, this uses the
[`continue-on-error`](https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idstepscontinue-on-error)
feature to allow some external partner tests to fail without failing the
entire build. **The downside** is that it is difficult to know when
these jobs are actually failing since they appear to have passed in
GitHub's UI. We will likely want to do something to address this in the
future.

Downsides

- GitHub Actions Workflows are still in beta so it's possible that there
could be breaking changes.