Enable `no-empty-pattern` rule for TypeScript files
Fixes two [no-empty-pattern](https://eslint.org/docs/rules/no-empty-pattern)
errors by disabling the rule for the violating lines.

I think this is actually a case where we're seeing false positives
because the rule was originally not implemented with TypeScript in mind,
so eslint is interpreting the type definition as an attempt to
destructure.

This needs at least an issue open in this repo but probably with
`@typescript-eslint` as well.