Fail CI build if `yarn.lock` is out of date
This adds the [`--frozen-lockfile`](https://yarnpkg.com/lang/en/docs/cli/install/#toc-yarn-install)
flag to our CI scripts where we are `yarn install`ing dependencies.

This flag will cause the script to exit with an error if the committed
`yarn.lock` is not up to date.

This makes sure we our `yarn.lock` file is always up to date.