fixed serialize example (#6816)
I saw in the [documentation](https://api.emberjs.com/ember-data/release/modules/@ember-data%2Fserializer) the example code for the `serialize` method and it shows the `snapshot.id` property as a function instead of a string. I [double checked it](https://github.com/emberjs/data/blob/master/packages/store/addon/-private/system/snapshot.js#L65) to be sure it is just a string :) I hope this is the right place to make the fix to the documentation.