[CHORE] improve comment bot (#6820)
- changes the comment-bot to update existing comments instead of deleting old comments
- fixes the bug that was preventing deletion of old comments
- fixes the performance analysis script to also post a comment correctly (it's been failing to do so although trying for some time)

You can see the effects of the updates (edit history on the pr comments) within this PR. Overall it seems like a good change, although I dislike how the comment time stamp remains the time of the very original comment instead of showing the time edited. This makes it harder to judge quickly if it is the most recent analysis, although checking the commit sha that is printed will verify.