[BUGFIX TESTS] matrix tests for fastboot & ie11 w/jquery & fetch (#6831)
resolves #6431

Bugfix: jQuery+fastboot

The added tests for fastboot revealed that fastboot+jQuery was broken because our computed property for getting the fastboot service wasn't setting and updating it's value properly leading to it always being false. This led us down the wrong code-path and ultimately meant that the url we provided to najax wasn't parsed correctly and would error.