Fix changelog for v3.16.2
This fixes the changelog by moving the changelog entries originally
intended for 3.16.1 to the section for 3.16.2.

For some more context, due to an unfortunate mishap by yours truly,
3.16.1 was never actually published to NPM. Instead, 3.16.2 was
published with what should have been 3.16.1. NPM does not support any
way to remedy this after the fact (and with good reason), but it would
still be nice for the changelog to be correct for anyone viewing it via
GitHub.