Breaking test for .load()ing sync belongsTo
I think it should be possible to call .load() on a sync belongsTo reference,
but if you do, you get an AssertionError:

Assertion Failed: You looked up the 'team' relationship on a 'person' with id 1
but some of the associated records were not loaded. Either make sure they are
all loaded together with the parent record, or specify that the relationship
is async (`belongsTo({ async: true })`)

This commit adds a breaking test to demonstrate this.