typofix `fetchRecord` -> `findRecord`
Description for `coalesceFindRequests` indicates that it will attempt to coalesce `fetchRecord` calls, but the relevant Ember Data concept is `findRecord`. If using the term `fetchRecord` here was intended to be a stand-in for `coalesceFindRequests` also affecting `belongsTo` relationships and `get`, it may be more clear to say something like
```
 By default the store will try to coalesce all `findRecord` and async relationship calls within the same runloop
```

or something similar. Either way it is a bit confusing to have `fetchRecord` here since it hints at a framework concept that does not have any definition.