move ember-auto-import from devDeps to deps
This is targeting the release-4.0 branch.

The idea here is to be consistent with what we did in ember-source -- within the 4.0 breaking change, we added ember-auto-import 2.0, which imposes the requirement that an app must have ember-auto-import >= 2 in order to use ember-source 4.0.

This PR does the same thing for the ember-data packages.

ember-auto-import 2 serves as a polyfill for embroider that allows addons to move to the new streamlined v2 format. By making it a requirement here at the breaking change, we're free to convert all the ember-data packages to v2 format within the 4.0 series without other breaking changes.