Invoke ember-auto-import analyzer (#8222)
Because the ember-data packages opt out of the standard ember-cli preprocessor registry, they won't get analyzed correctly by ember-auto-import.

Instead, they're forced to manually configure the analyzer. That's what this PR does.

The `ember-auto-import/babel-plugin` that was being added before was doing nothing. As documented, it exists to support dynamic import, which ember-data doesn't use. It's unrelated to the import analysis.