Fix check for new records in JSONAPISerializer.serializeHasMany (#8794)
Use `!Snapshot.isNew` instead of `Snapshot.record && !Snapshot.record.isNew` to avoid the access to `record`. Accessing `record` creates an entry in `instanceCache.record` which causes the `hasRecord` checks in `Snapshot` to return true. This results in an error when serializing associated records via hasMany that have not been loaded.